src/web/requirements.txt
astroid==1.4.8
colorama==0.3.7
coverage==4.2
Django==1.10
django-ace==1.0.2
django-cleanup==0.4.2
django-filter==0.14.0
django-jenkins==0.19.0
django-markdown-deux==1.0.5
django-nested-inline==0.3.6
django-pagedown==0.1.1
djangorestframework==3.5.2
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.6
markdown2==2.3.1
mccabe==0.5.2
pep8==1.7.0
pyflakes==1.3.0
pylint==1.6.4
six==1.10.0
virtualenv==15.0.3
wrapt==1.10.8
jsonfield
django-sortedm2m
sorl-thumbnail